refactor(core): ファイル分割
